+++ ruby/lib/svix/models/endpoint_transformation_in.rb
@@ -3,22 +3,24 @@
 require "json"
 
 module Svix
-  class ExpungAllContentsOut
-    attr_accessor :id
-    attr_accessor :status
-    attr_accessor :task
+  class EndpointTransformationIn
+    attr_accessor :code
+    attr_accessor :enabled
 
-    ALL_FIELD ||= ["id", "status", "task"].freeze
+    ALL_FIELD ||= ["code", "enabled"].freeze
     private_constant :ALL_FIELD
 
     def initialize(attributes = {})
       unless attributes.is_a?(Hash)
-        fail(ArgumentError, "The input argument (attributes) must be a hash in `Svix::ExpungAllContentsOut` new method")
+        fail(
+          ArgumentError,
+          "The input argument (attributes) must be a hash in `Svix::EndpointTransformationIn` new method"
+        )
       end
 
       attributes.each do |k, v|
         unless ALL_FIELD.include?(k.to_s)
-          fail(ArgumentError, "The field #{k} is not part of Svix::ExpungAllContentsOut")
+          fail(ArgumentError, "The field #{k} is not part of Svix::EndpointTransformationIn")
         end
 
         instance_variable_set("@#{k}", v)
@@ -29,17 +31,15 @@ module Svix
     def self.deserialize(attributes = {})
       attributes = attributes.transform_keys(&:to_s)
       attrs = Hash.new
-      attrs["id"] = attributes["id"]
-      attrs["status"] = Svix::BackgroundTaskStatus.deserialize(attributes["status"])
-      attrs["task"] = Svix::BackgroundTaskType.deserialize(attributes["task"])
+      attrs["code"] = attributes["code"]
+      attrs["enabled"] = attributes["enabled"]
       new(attrs)
     end
 
     def serialize
       out = Hash.new
-      out["id"] = Svix::serialize_primitive(@id) if @id
-      out["status"] = @status.serialize if @status
-      out["task"] = @task.serialize if @task
+      out["code"] = Svix::serialize_primitive(@code) if @code
+      out["enabled"] = Svix::serialize_primitive(@enabled) if @enabled
       out
     end