+++ rust/src/api/endpoint.rs
@@ -1,3 +1,4 @@
+// this file is @generated
 use crate::{error::Result, models::*, Configuration};
 
 #[derive(Default)]
@@ -382,6 +383,26 @@ impl<'a> Endpoint<'a> {
         .with_path_param("app_id", app_id)
         .with_path_param("endpoint_id", endpoint_id)
         .with_body_param(endpoint_transformation_patch)
+        .returns_nothing()
+        .execute(self.cfg)
+        .await
+    }
+
+    /// This operation was renamed to `set-transformation`.
+    #[deprecated]
+    pub async fn transformation_partial_update(
+        &self,
+        app_id: String,
+        endpoint_id: String,
+        endpoint_transformation_in: EndpointTransformationIn,
+    ) -> Result<()> {
+        crate::request::Request::new(
+            http1::Method::PATCH,
+            "/api/v1/app/{app_id}/endpoint/{endpoint_id}/transformation",
+        )
+        .with_path_param("app_id", app_id)
+        .with_path_param("endpoint_id", endpoint_id)
+        .with_body_param(endpoint_transformation_in)
         .returns_nothing()
         .execute(self.cfg)
         .await